Customizable

Description

Este es el modulo que se expone al usuario para personalizar el comportamiento de el programa. En esencia los pasos para crear un estilo de carta nueva es crear un tipo nuevo.

newtype AnkItemCarta = AnkItemCarta { ankItemCarta :: ItemCarta}

luego hacer este tipo de dato una instancia de la clase ToHtml.

GeneralizedNewtypeDeriving

Documentation

priceOne' :: (Monad m, PricedOnce a) => a -> HtmlT m ()

priceTwo' :: (Monad m, PricedTwice a) => a -> HtmlT m ()

priceThree' :: (Monad m, PricedThree a) => a -> HtmlT m ()

itemsDescription' :: (Monad m, NamedItem a) => a -> HtmlT m ()

itemsName' :: (Monad m, NamedItem a) => a -> HtmlT m ()

priceFour' :: Monad m => Item4 -> HtmlT m ()

formatPrice :: String -> String

formatItem :: Item -> Item

itemStyle :: Monad m => Item -> HtmlT m ()

item2Style :: Monad m => Item2 -> HtmlT m ()

item3Style :: Monad m => Item3 -> HtmlT m ()

newtype AnkItemCarta

Constructors

AnkItemCarta

Fields

ankItemCarta :: ItemCarta

newtype MoltoItemCarta

Constructors

MoltoItemCarta

Fields

moltoItemCarta :: ItemCarta

newtype BrasasItemCarta

Constructors

BrasasItemCarta

Fields

brasasItemCarta :: ItemCarta

newtype Sushi7ItemCarta

Constructors

Sushi7ItemCarta

Fields

sushi7ItemCarta :: ItemCarta

newtype PicoteoItemCarta

Constructors

PicoteoItemCarta

Fields

picoteoItemCarta :: ItemCarta

newtype WajacaItemCarta

Constructors

WajacaItemCarta

Fields

wajacaItemCarta :: ItemCarta

newtype VillageItemCarta

Constructors

VillageItemCarta

Fields

villageItemCarta :: ItemCarta

newtype FederalRibsItemCarta

Constructors

FederalRibsItemCarta

Fields

federalRibsItemCarta :: ItemCarta